requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
gunicorn==20.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2